Change Folder Names
Makes more sense
