+++ src/main/java/SMARTron/GUIMiddleware/PseudoStudent.java
