+++ src/main/java/SMARTron/Scanner/Student.java
