+++ src/main/java/SMARTron/Scanner/Tools.java
