+++ src/main/java/SMARTron/Scanner/Utilities.java
